Normalize decimals coming out of the Gemini importer
For some reason exponentiation seems to be adding a lot of extra
precision. The Decimal class assumes that the number of significant
digits in the input should be preserved, but something about the way we
exponentiate seems to add even more context digits that needed.

normalize() strips any trailing zeros that are precision padding.